fix: extraction de la logique dans reportedCompany.service.ts
